cloud-agent/requirements.txt
# Cloud Agent Requirements - Platinum Tier

# Base requirements (from main requirements.txt)
watchdog==3.0.0
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0

# Gmail API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# System monitoring
psutil==5.9.6

# Git operations (usually pre-installed, but just in case)
GitPython==3.1.40

# Logging and monitoring
python-json-logger==2.0.7
